+++ Locales/enUS.lua
@@ -89,6 +89,16 @@ select(2, ...).L = setmetatable({
         <h2>If there are any issues after an update, check through all code snippets first.</h2>
         <br/>
 
+        <h1>r262-release (Sep 18, 2025, 18:50 GMT+8)</h1>
+        <p>* Updated MoP debuffs (#417, thanks Ekkles).</p>
+        <p>* Added missing paladin click-casting spell (#389, thanks gpejski).</p>
+        <p>* Partial aura updates (#413, thanks cont1nuity).</p>
+        <p>* (Retail) Refactored aura update logic and cache handling, slightly improved performance.</p>
+        <p>* Added option to hide Blizzard Raid Manager.</p>
+        <p>* Added "Cast By" for custom debuff indicators.</p>
+        <p>* Updated TWW debuffs and targeted spells.</p>
+        <br/>
+
         <h1>r261-release (Aug 22, 2025, 20:40 GMT+8)</h1>
         <p>* Updated MoP debuffs (thanks Ekkles).</p>
         <p>* Updated TWW debuffs.</p>